Preparatory commit prior to PreJagati Sprint.
This is a dump of files from the Mezzanine and they are frozen until
this package comes online. This is not intended to be functional or
even correct yet. This will be mutated until it works as intended
and will be the first Jagati package truly cut out of the Mezzanine.