Fixed a bug where we'd write outside the bounds of a provided array.
Credit where credit was due, MSVC caught a buffer overflow for us.  We were assuming two writes on every loop where that isn't always the case.  Any time a Base64 string ends with "==" then only one character is being written.  With the old code we'd write two willy nilly and check if we'd need to write a third which would cause us to write an additional character (a null terminator) 1 byte outside the valid bounds of the buffer.