Attempt to fix ManagedArrayTests
A few different sources including Gcc docs and mailing list indicate that
Link-Time Optimization and the -Wstrict-overflow warning can be spuriously
triggered in certain cases involving pointer arithmetic nested between
multiple if statments. All of the ManagedArray is typesafe and operates in
terms of iterators and indexes. It seems that in Release builds the type
are more completely optimized out and the pure pointer math remains. This
then goes through optimizations during later phases of the build (like
linking) and since the types are gone it is unknown if the pointer math
might overflow triggering UB.

It seems that more metadata remains if the files involved are smaller.
After a few dozen assertions the error Related to strict overflow
appeared. So this splits the test file in two.

This is a dirty hack.