Exception cross project fix part1 (#65)
* This is a rough draft using the new Jagati exceptions feature.

* Changed copyright, updated Jagati to 0.29.0 and added minimal tests

* Adds exception support to the Foundation

* Update to Jagati 0.29.0

* Fixed PR comments

* A trivial change to trigger a Jenkins build

* This should allow the Foundation to build with the older jagati and the newer one that will require an exception initilization.

* Lengthened Base64 test to make it less unstable and put correct names in Jenkinsfile

* Fixing capitilized MSVC in Jenkinsfile

* Attempt to fix MSVC build in Jenkins

* Testing parallel builds with same ageent for a completely unrelated task

* Updates to JenkinsFile

* This should fix the Jenkinsfile and use the new multi-repo exceptions.

* Fixing some tests in Jenkins and isolated odd failure case. Testing CI again.

* Fixing no-var-tracking on clang compilers.

* Updating to Jagati 0.30.1

* Attempt to fix ManagedArrayTests

A few different sources including Gcc docs and mailing list indicate that
Link-Time Optimization and the -Wstrict-overflow warning can be spuriously
triggered in certain cases involving pointer arithmetic nested between
multiple if statments. All of the ManagedArray is typesafe and operates in
terms of iterators and indexes. It seems that in Release builds the type
are more completely optimized out and the pure pointer math remains. This
then goes through optimizations during later phases of the build (like
linking) and since the types are gone it is unknown if the pointer math
might overflow triggering UB.

It seems that more metadata remains if the files involved are smaller.
After a few dozen assertions the error Related to strict overflow
appeared. So this splits the test file in two.

This is a dirty hack.